Separate out the collections
